webpage/backend/requirements.txt
fastapi==0.118.2
uvicorn[standard]==0.37.0
elasticsearch[async]==8.19.1
pillow==11.3.0
pydantic==2.12.0
httpx==0.28.1
PyJWT[crypto]==2.10.1
